Update to work with new custom draft system.
