+++ rollup.config.js
@@ -1,5 +1,5 @@
 export default {
-  input: ['src/util/Filter.js', 'src/util/Card.js'],
+  input: ['src/util/Filter.js', 'src/util/Card.js', 'src/util/draftutil.js', 'src/util/Util.js'],
   output: {
     dir: 'dist/util',
     format: 'cjs',
